% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_lysiphen.R
\name{lysiphen}
\alias{lysiphen}
\title{Perform LysipheN data analysis and vis}
\usage{
lysiphen(data_path, file_ext = "csv", threshold = 100)
}
\arguments{
\item{data_path}{String with the path to locate the LysipheN file(s).}

\item{file_ext}{csv by default. Also xls or xlsx available.}

\item{threshold}{Integer of length 1 with the maximum weight change allowed (in grams). Any weight lower or higher than this offset will be detected as outlier and imputed by the immediate previous one.}
}
\value{
A list
}
\description{
The function loads the data generated by the LysipheN devices; detects outliers based on the defined threshold; calculates transpiration per device and per crop; and creates plots
}
\examples{
lysiphen(data_path = getwd(), file_ext = "csv", threshold = 100)

}
